🎨 Fusionne les boutons "Passer →" et "Suivant →"
