+++ mon-entreprise/source/components/conversation/Conversation.tsx
@@ -1,22 +1,20 @@
 import { goToQuestion, validateStepWithValue } from 'Actions/actions'
-import QuickLinks from 'Components/QuickLinks'
 import RuleInput from 'Components/conversation/RuleInput'
-import React, { useContext, useEffect } from 'react'
+import QuickLinks from 'Components/QuickLinks'
+import * as Animate from 'Components/ui/animate'
+import { EngineContext } from 'Components/utils/EngineContext'
+import { useNextQuestions } from 'Components/utils/useNextQuestion'
+import { default as React, useContext, useEffect } from 'react'
 import emoji from 'react-easy-emoji'
 import { Trans } from 'react-i18next'
 import { useDispatch, useSelector } from 'react-redux'
-import { RootState } from 'Reducers/rootReducer'
-import * as Animate from 'Components/ui/animate'
+import {
+	answeredQuestionsSelector,
+	situationSelector
+} from 'Selectors/simulationSelectors'
 import Aide from './Aide'
 import './conversation.css'
 import FormDecorator from './FormDecorator'
-import { useNextQuestions } from 'Components/utils/useNextQuestion'
-import { EngineContext } from 'Components/utils/EngineContext'
-import PreviousAnswers from 'sites/mon-entreprise.fr/pages/Créer/GuideStatut/PreviousAnswers'
-import {
-	answeredQuestionsSelector,
-	currentQuestionSelector
-} from 'Selectors/simulationSelectors'
 
 export type ConversationProps = {
 	customEndMessages?: React.ReactNode
@@ -26,6 +24,8 @@ export default function Conversation({ customEndMessages }: ConversationProps) {
 	const dispatch = useDispatch()
 	const rules = useContext(EngineContext).getParsedRules()
 	const currentQuestion = useNextQuestions()[0]
+	const situation = useSelector(situationSelector)
+	const currentQuestionIsAnswered = !!situation[currentQuestion]
 
 	const previousAnswers = useSelector(answeredQuestionsSelector)
 	const setDefault = () =>
@@ -42,6 +42,13 @@ export default function Conversation({ customEndMessages }: ConversationProps) {
 			setDefault()
 		}
 	}
+	const submit = source =>
+		dispatch({
+			type: 'STEP_ACTION',
+			name: 'fold',
+			step: currentQuestion,
+			source
+		})
 	const DecoratedInputComponent = FormDecorator(RuleInput)
 
 	return currentQuestion ? (
@@ -62,12 +69,16 @@ export default function Conversation({ customEndMessages }: ConversationProps) {
 							</button>
 						</>
 					)}
-					<button
-						onClick={setDefault}
-						className="ui__ simple small push-right button"
-					>
-						<Trans>Passer</Trans> →
-					</button>
+					{currentQuestionIsAnswered ? (
+						<SendButton onSubmit={submit} />
+					) : (
+						<button
+							onClick={setDefault}
+							className="ui__ simple small push-right button"
+						>
+							<Trans>Passer</Trans> →
+						</button>
+					)}
 				</div>
 			</div>
 			<QuickLinks />
@@ -90,5 +101,35 @@ export default function Conversation({ customEndMessages }: ConversationProps) {
 				)}
 			</p>
 		</div>
+	)
+}
+
+type SendButtonProps = {
+	onSubmit: (cause: string) => void
+}
+
+function SendButton({ onSubmit }: SendButtonProps) {
+	useEffect(() => {
+		const handleKeyDown = ({ key }: KeyboardEvent) => {
+			if (key !== 'Enter') return
+			onSubmit('enter')
+		}
+
+		window.addEventListener('keydown', handleKeyDown)
+		return () => {
+			window.removeEventListener('keydown', handleKeyDown)
+		}
+	}, [onSubmit])
+
+	return (
+		<button
+			className="ui__ plain button "
+			css="margin-left: 1.2rem"
+			onClick={() => onSubmit('accept')}
+		>
+			<span className="text">
+				<Trans>Suivant</Trans> →
+			</span>
+		</button>
 	)
 }
