+++ mon-entreprise/source/components/conversation/DateInput.tsx
@@ -3,7 +3,6 @@ import { Rule } from 'publicodes'
 import React, { useCallback, useMemo } from 'react'
 import styled from 'styled-components'
 import InputSuggestions from './InputSuggestions'
-import SendButton from './SendButton'
 
 type DateInputProps = {
 	onChange: RuleInputProps['onChange']
@@ -57,7 +56,6 @@ export default function DateInput({
 					onChange={handleDateChange}
 				/>
 			</div>
-			{onSubmit && <SendButton disabled={!dateValue} onSubmit={onSubmit} />}
 		</div>
 	)
 }
