+++ mon-entreprise/source/components/conversation/Question.tsx
@@ -5,7 +5,6 @@ import React, { useCallback, useContext } from 'react'
 import emoji from 'react-easy-emoji'
 import { Trans } from 'react-i18next'
 import { ExplicableRule } from './Explicable'
-import SendButton from './SendButton'
 
 /* Ceci est une saisie de type "radio" : l'utilisateur choisit une réponse dans
 	une liste, ou une liste de listes. Les données @choices sont un arbre de type:
@@ -115,7 +114,6 @@ export default function Question({
 			css="margin-top: 0.6rem; display: flex; align-items: center; flex-wrap: wrap;"
 		>
 			{choiceElements}
-			{onSubmit && <SendButton disabled={!currentValue} onSubmit={onSubmit} />}
 		</div>
 	)
 }
